pyproject.toml
[tool.poetry]
name = "dbsync"
version = "0.1.0"
description = "sync different databases"
authors = ["Dongook Son <60206749+donny-son@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
python-dotenv = "^0.21.0"
sh = "^1.14.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
jedi-language-server = "^0.37.0"
pytest = "^7.1.3"
pre-commit = "^2.20.0"
debugpy = "^1.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
